Move entity models into app
