--- entity/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "spice_entity"
-version = "0.1.0"
-edition = "2024"
-
-[lib]
-path = "src/mod.rs"
-
-[dependencies]
-axum = { version = "0.8.4", features = ["multipart"] }
-chrono = { version = "0.4.40", features = ["serde"] }
-# crudcrate = "0.4.1"
-crudcrate = { path = "../../crudcrate" }
-hyper = "1.6.0"
-sea-orm = { version = "1.1.14", features = [
-    "macros",
-    "with-uuid",
-    "with-chrono",
-    "with-json",
-    "with-rust_decimal",
-], default-features = false }
-serde = { version = "1.0.217", features = ["derive"] }
-serde_json = "1.0.137"
-utoipa = { version = "5.3.1", features = [
-    "axum_extras",
-    "uuid",
-    "chrono",
-    "decimal",
-] }
-utoipa-axum = "0.2.0"
-utoipa-scalar = { version = "0.3.0", features = ["axum"] }
-uuid = { version = "1.13.2", features = ["serde", "v4", "fast-rng"] }
