--- entity/src/mod.rs
@@ -1,21 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
-
-pub mod prelude;
-
-pub mod experiments;
-pub mod locations;
-pub mod projects;
-pub mod regions;
-pub mod s3_assets;
-pub mod samples;
-pub mod sea_orm_active_enums;
-pub mod spatial_ref_sys;
-pub mod temperature_readings;
-pub mod tray_configuration_assignments;
-pub mod tray_configurations;
-pub mod trays;
-pub mod treatments;
-pub mod well_phase_transitions;
-pub mod wells;
-
-pub use sea_orm_active_enums::SampleType;
