--- entity/src/prelude.rs
@@ -1,16 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
-
-pub use super::experiments::Entity as Experiments;
-pub use super::locations::Entity as Locations;
-pub use super::projects::Entity as Projects;
-pub use super::regions::Entity as Regions;
-pub use super::s3_assets::Entity as S3Assets;
-pub use super::samples::Entity as Samples;
-pub use super::spatial_ref_sys::Entity as SpatialRefSys;
-pub use super::temperature_readings::Entity as TemperatureReadings;
-pub use super::tray_configuration_assignments::Entity as TrayConfigurationAssignments;
-pub use super::tray_configurations::Entity as TrayConfigurations;
-pub use super::trays::Entity as Trays;
-pub use super::treatments::Entity as Treatments;
-pub use super::well_phase_transitions::Entity as WellPhaseTransitions;
-pub use super::wells::Entity as Wells;
