--- entity/src/s3_assets.rs
@@ -1,46 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "s3_assets")]
-pub struct Model {
-    pub experiment_id: Option<Uuid>,
-    #[sea_orm(column_type = "Text")]
-    pub original_filename: String,
-    #[sea_orm(column_type = "Text", unique)]
-    pub s3_key: String,
-    pub size_bytes: Option<i64>,
-    #[sea_orm(column_type = "Text", nullable)]
-    pub uploaded_by: Option<String>,
-    pub uploaded_at: DateTimeWithTimeZone,
-    pub is_deleted: bool,
-    pub created_at: DateTimeWithTimeZone,
-    pub last_updated: DateTimeWithTimeZone,
-    #[sea_orm(column_type = "Text")]
-    pub r#type: String,
-    #[sea_orm(column_type = "Text", nullable)]
-    pub role: Option<String>,
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: Uuid,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::experiments::Entity",
-        from = "Column::ExperimentId",
-        to = "super::experiments::Column::Id",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    Experiments,
-}
-
-impl Related<super::experiments::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Experiments.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
