--- entity/src/sea_orm_active_enums.rs
@@ -1,39 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
-
-use sea_orm::entity::prelude::*;
-use serde::{Deserialize, Serialize};
-use utoipa::ToSchema;
-
-#[derive(
-    Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum, ToSchema, Serialize, Deserialize,
-)]
-#[sea_orm(rs_type = "String", db_type = "Enum", enum_name = "sample_type")]
-pub enum SampleType {
-    #[sea_orm(string_value = "bulk")]
-    #[serde(rename = "bulk")]
-    Bulk,
-    #[sea_orm(string_value = "filter")]
-    #[serde(rename = "filter")]
-    Filter,
-    #[sea_orm(string_value = "procedural_blank")]
-    #[serde(rename = "procedural_blank")]
-    ProceduralBlank,
-    #[sea_orm(string_value = "pure_water")]
-    #[serde(rename = "pure_water")]
-    PureWater,
-}
-#[derive(
-    Debug, Clone, PartialEq, Eq, EnumIter, DeriveActiveEnum, ToSchema, Serialize, Deserialize,
-)]
-#[sea_orm(rs_type = "String", db_type = "Enum", enum_name = "treatment_name")]
-pub enum TreatmentName {
-    #[sea_orm(string_value = "none")]
-    #[serde(rename = "none")]
-    None,
-    #[sea_orm(string_value = "heat")]
-    #[serde(rename = "heat")]
-    Heat,
-    #[sea_orm(string_value = "h2o2")]
-    #[serde(rename = "h2o2")]
-    H2o2,
-}
