--- entity/src/spatial_ref_sys.rs
@@ -1,19 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "spatial_ref_sys")]
-pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub srid: i32,
-    pub auth_name: Option<String>,
-    pub auth_srid: Option<i32>,
-    pub srtext: Option<String>,
-    pub proj4text: Option<String>,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {}
-
-impl ActiveModelBehavior for ActiveModel {}
