--- entity/src/treatments.rs
@@ -1,47 +1,0 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
-
-use super::sea_orm_active_enums::TreatmentName;
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "treatments")]
-pub struct Model {
-    #[sea_orm(column_type = "Text", nullable)]
-    pub notes: Option<String>,
-    pub sample_id: Option<Uuid>,
-    pub last_updated: DateTimeWithTimeZone,
-    pub created_at: DateTimeWithTimeZone,
-    #[sea_orm(column_type = "Decimal(Some((16, 10)))", nullable)]
-    pub enzyme_volume_litres: Option<Decimal>,
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub id: Uuid,
-    pub name: TreatmentName,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(has_many = "super::regions::Entity")]
-    Regions,
-    #[sea_orm(
-        belongs_to = "super::samples::Entity",
-        from = "Column::SampleId",
-        to = "super::samples::Column::Id",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    Samples,
-}
-
-impl Related<super::regions::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Regions.def()
-    }
-}
-
-impl Related<super::samples::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Samples.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
