+++ src/main.rs
@@ -4,6 +4,7 @@ mod external;
 mod routes;
 mod services;
 
+
 use crate::config::Config;
 use migration::{Migrator, MigratorTrait};
 use sea_orm::{Database, DatabaseConnection};
