+++ src/routes/experiments/mod.rs
@@ -1,6 +1,8 @@
 pub mod excel_upload;
 pub mod models;
+pub mod phase_transitions;
 pub mod services;
+pub mod temperatures;
 #[cfg(test)]
 mod tests;
 pub mod views;
