+++ src/routes/projects/mod.rs
@@ -1,4 +1,4 @@
-pub mod views;
-
+pub mod models;
 #[cfg(test)]
 pub mod tests;
+pub mod views;
