+++ src/routes/projects/models.rs
@@ -1,5 +1,3 @@
-//! `SeaORM` Entity, @generated by sea-orm-codegen 1.1.8
-
 use super::locations::Location;
 use chrono::{DateTime, Utc};
 use crudcrate::{CRUDResource, EntityToModels};
@@ -13,7 +11,7 @@ use sea_orm::entity::prelude::*;
     name_singular = "project",
     name_plural = "projects",
     description = "Projects provide a way to organise locations hierarchically. Each project can contain multiple locations and provides visual organization through color coding.",
-    fn_get_one = get_one
+    fn_get_one = get_one,
 )]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
@@ -23,16 +21,16 @@ pub struct Model {
     #[crudcrate(sortable, filterable, fulltext)]
     pub name: String,
     #[sea_orm(column_type = "Text", nullable)]
-    #[crudcrate(sortable, filterable, fulltext)]
+    #[crudcrate(sortable, filterable, fulltext, list_model = false)]
     pub note: Option<String>,
     #[crudcrate(sortable, filterable, fulltext)]
     pub colour: Option<String>,
-    #[crudcrate(update_model = false, create_model = false, on_create = chrono::Utc::now(), sortable)]
+    #[crudcrate(update_model = false, create_model = false, on_create = chrono::Utc::now(), sortable, list_model=false)]
     pub created_at: DateTime<Utc>,
-    #[crudcrate(update_model = false, create_model = false, on_update = chrono::Utc::now(), on_create = chrono::Utc::now(), sortable)]
+    #[crudcrate(update_model = false, create_model = false, on_update = chrono::Utc::now(), on_create = chrono::Utc::now(), sortable, list_model=false)]
     pub last_updated: DateTime<Utc>,
     #[sea_orm(ignore)]
-    #[crudcrate(non_db_attr = true, default = vec![])]
+    #[crudcrate(non_db_attr = true, default = vec![], list_model = false)]
     pub locations: Vec<Location>,
 }
 
@@ -51,26 +49,20 @@ impl Related<super::locations::Entity> for Entity {
 impl ActiveModelBehavior for ActiveModel {}
 
 async fn get_one(db: &DatabaseConnection, id: Uuid) -> Result<Project, DbErr> {
-    // Custom get one, to collect location objects
-    let model: Model = match Entity::find_by_id(id).one(db).await? {
-        Some(model) => model,
-        None => {
-            return Err(DbErr::RecordNotFound(format!(
-                "{} not found",
-                Project::RESOURCE_NAME_SINGULAR
-            )));
-        }
-    };
+    let model = Entity::find_by_id(id)
+        .one(db)
+        .await?
+        .ok_or_else(|| DbErr::RecordNotFound("Project not found".to_string()))?;
 
-    let locations = model.find_related(super::locations::Entity).all(db).await?;
-
-    let location_objs: Vec<super::locations::Location> = locations
+    let locations: Vec<super::locations::Location> = model
+        .find_related(super::locations::Entity)
+        .all(db)
+        .await?
         .into_iter()
-        .map(std::convert::Into::into)
+        .map(Into::into)
         .collect();
 
     let mut project: Project = model.into();
-    project.locations = location_objs;
-
+    project.locations = locations;
     Ok(project)
 }
