+++ src/routes/trays/configuration_assignments/mod.rs
@@ -1,0 +1,1 @@
+pub mod models;
