+++ src/routes/trays/configuration_assignments/models.rs
