+++ src/routes/trays/mod.rs
@@ -1,5 +1,9 @@
+pub mod configuration_assignments;
+pub mod configurations;
 pub mod models;
+pub mod regions;
 pub mod services;
 #[cfg(test)]
 mod tests;
 pub mod views;
+pub mod wells;
